refactor(nativeBatching): colocate the hybrid schema builder with its only consumer
