+++ src/services/destination/nativeBatching/batchDestination.ts
@@ -19,7 +19,7 @@ export type {
   ExtractConnectionConfig,
 } from './types';
 export { BodyFormat, parseSizeToBytes } from './types';
-export { makeRouterInputSchema, makeHybridInputSchema } from './inputSchema';
+export { makeRouterInputSchema } from './inputSchema';
 export { ChunkBatchStrategy } from './chunkBatchStrategy';
 export { CustomBatchStrategy } from './customBatchStrategy';
